requirements.txt
# API and Web UI
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.4.2
python-dotenv>=1.0.0

# LangChain Components
langchain>=0.0.335
langchain_community>=0.0.6
langchain-huggingface>=0.0.2

# Vector Database and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# NLP and ML
transformers>=4.35.0
torch>=2.1.0
google-generativeai>=0.3.1

# Utilities
requests>=2.31.0
pandas>=2.1.1
numpy>=1.26.0
pillow>=10.0.1 